requirements.txt
fastapi
openai
python-dotenv
phospho